+++ NEWS.md
@@ -1,6 +1,7 @@
 # teamcolors 0.0.2
 
 * Fixed Los Angeles of Anaheim division
+* Updated New York Jets color (https://github.com/beanumber/teamcolors/issues/7)
 
 # teamcolors 0.0.1