+++ NEWS.md
@@ -2,6 +2,10 @@
 
 * Fixed Los Angeles of Anaheim division
 * Updated New York Jets color (https://github.com/beanumber/teamcolors/issues/7)
+* Added `show_team_col()` function
+* Added `ggplot2` functionality for `scale_color/fill_teams()`
+* Added color palette functionality through named vectors
+* Added links to logos
 
 # teamcolors 0.0.1